requirements.txt
psycopg[binary]>=3.1
curl_cffi>=0.7
playwright>=1.40
python-dotenv>=1.0
fastapi>=0.115
uvicorn[standard]>=0.30
